Removed compatibility with old versions
* We support just latest versions:
  * Spigot 1.16.5
  * w/e w/g 7.X
