+++ areashop-interface/pom.xml
@@ -27,7 +27,7 @@
         <dependency>
             <groupId>com.sk89q.worldguard</groupId>
             <artifactId>worldguard-bukkit</artifactId>
-            <version>7.0.2-SNAPSHOT</version>
+            <version>7.0.4-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
     </dependencies>
