+++ areashop-worldguard-7/pom.xml
@@ -28,14 +28,14 @@
         <dependency>
             <groupId>com.sk89q.worldguard</groupId>
             <artifactId>worldguard-bukkit</artifactId>
-            <version>7.0.2-SNAPSHOT</version>
+            <version>7.0.4-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
 
         <dependency>
             <groupId>me.wiefferink</groupId>
             <artifactId>areashop-interface</artifactId>
-            <version>latest</version>
+            <version>2.6.0-SNAPSHOT</version>
             <type>jar</type>
         </dependency>
     </dependencies>
